Deduplicate drop database, fix some bugs, rewrite RationalTable, use Double over Rational for more precision

+++ src/main/kotlin/api/api/predef/ext/RationalExt.kt
@@ -5,4 +5,4 @@ import io.luna.util.Rational
 /**
  * Creates a [Rational] instance.
  */
-infix fun Int.of(d: Int) = Rational(this.toLong(), d.toLong())
+infix fun Int.of(d: Int) = Rational(this.toLong(), d.toLong()).toDouble()
